+++ packages/core-js-compat/src/data.js
@@ -1266,6 +1266,7 @@ const data = {
   // TODO: Remove from `core-js@4`
   'esnext.aggregate-error': null,
   'esnext.array.at': {
+    chrome: '92',
   },
   'esnext.array.filter-out': {
   },
@@ -1517,6 +1518,7 @@ const data = {
   'esnext.symbol.replace-all': {
   },
   'esnext.typed-array.at': {
+    chrome: '92',
   },
   'esnext.typed-array.filter-out': {
   },