+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 ##### Unreleased
+- Nothing
+
+##### 3.14.0 - 2021.06.05
 - Added polyfill of stable sort in `{ Array, %TypedArray% }.prototype.sort`, [#769](https://github.com/zloirock/core-js/issues/769)
 - Fixed `Safari` 14.0- `%TypedArray%.prototype.sort` validation of arguments bug
 - `.at` marked as supported from V8 9.2